services/ingress-bridge/requirements.txt
nats-py>=2.7.0
confluent-kafka>=2.5.0
zstandard>=0.22.0
protobuf>=4.25.0
grpcio-tools>=1.60.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0
prometheus-client>=0.19.0
minio>=7.2.0
